refactor(decomposition): removed unused import
